refactor: moved meade files in a subfolder
