+++ src/core/meade/MeadeParserInit.cpp
@@ -3,7 +3,7 @@
  * @brief Init-family (`:I...`) dispatcher for the Meade LX200 parser.
  */
 
-#include "core/MeadeParser.hpp"
+#include "MeadeParser.hpp"
 
 namespace oat
 {
