+++ src/core/meade/MeadeParserSet.cpp
@@ -3,8 +3,8 @@
  * @brief Set-family (`:S...`) dispatcher for the Meade LX200 parser.
  */
 
-#include "core/MeadeParser.hpp"
-#include "core/MeadeParserHelpers.hpp"
+#include "MeadeParser.hpp"
+#include "MeadeParserHelpers.hpp"
 
 #include <stddef.h>
 #include <stdint.h>
