+++ src/core/meade/MeadeParserSlewRate.cpp
@@ -3,7 +3,7 @@
  * @brief SetSlewRate-family (`:R...`) dispatcher for the Meade LX200 parser.
  */
 
-#include "core/MeadeParser.hpp"
+#include "MeadeParser.hpp"
 
 namespace oat
 {
