+++ src/core/meade/MeadeParserSyncControl.cpp
@@ -3,8 +3,8 @@
  * @brief SyncControl-family (`:C...`) dispatcher for the Meade LX200 parser.
  */
 
-#include "core/MeadeParser.hpp"
-#include "core/MeadeParserHelpers.hpp"
+#include "MeadeParser.hpp"
+#include "MeadeParserHelpers.hpp"
 
 namespace oat
 {
