+++ src/core/meade/MeadeProtocol.hpp
