+++ unit_tests/test_meade/test_MeadeExtra.cpp
@@ -12,7 +12,7 @@
 
 #include <unity.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
