+++ unit_tests/test_meade/test_MeadeGet.cpp
@@ -9,7 +9,7 @@
 
 #include <string.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
