+++ unit_tests/test_meade/test_MeadeGps.cpp
@@ -8,7 +8,7 @@
 
 #include <unity.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
