+++ unit_tests/test_meade/test_MeadeHome.cpp
@@ -6,7 +6,7 @@
 
 #include <unity.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
