+++ unit_tests/test_meade/test_MeadeInit.cpp
@@ -4,7 +4,7 @@
 
 #include <unity.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
