+++ unit_tests/test_meade/test_MeadeQuit.cpp
@@ -8,7 +8,7 @@
 
 #include <string.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
