+++ unit_tests/test_meade/test_MeadeSet.cpp
@@ -10,7 +10,7 @@
 
 #include <string.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
