+++ unit_tests/test_meade/test_MeadeSlewRate.cpp
@@ -5,7 +5,7 @@
 
 #include <unity.h>
 
-#include "core/MeadeParser.hpp"
+#include "core/meade/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
